refactor: move auth related use cases to a package
