+++ src/infra/dependencies.py
@@ -1,16 +1,16 @@
 from typing import Annotated
 
 from fastapi import Depends, HTTPException
-from fastapi.security import OAuth2PasswordBearer
 from fastapi import status
+from fastapi.security import OAuth2PasswordBearer
 from jose import jwt, JWTError
 
 from src.adapters.repos.auth import AuthRepo
 from src.adapters.repos.vehicle import VehicleRepo
-from src.domain.use_cases.login import LoginUseCase
+from src.domain.use_cases.auth.login import LoginUseCase
+from src.domain.use_cases.auth.read_user_info import ReadUserInfoUseCase
+from src.domain.use_cases.auth.signup import SignupUseCase
 from src.domain.use_cases.publish_vehicle import PublishVehicleUseCase
-from src.domain.use_cases.read_user_info import ReadUserInfoUseCase
-from src.domain.use_cases.signup import SignupUseCase
 from src.infra import config
 from src.infra.db import Database, AiohttpHttpNode
